topdown/copypropagation: Move unionfind to new file
Just moving the code, no other changes.
